refactor: centralize orphan cleanup helpers and tighten cleaned semantics
- Move cleanupOrphanPods and cleanupOrphanPodSets to utils.go alongside
  isOwnedByRoleSet so all orphan-cleanup helpers live together.
- Track cleaned only when at least one Delete actually succeeds, so the
  returned bool reflects real cluster mutation rather than mere detection.
